Fix incomplete gender support in two logentry messages
The first three parameters for these messages come from core's
LogFormatter::getMessageParameters. Parameter number 4 comes from
SpecialLogLogger::log in this codebase. It's a raw username and fine
to be used for GENDER:… because of this. No change to the code is
needed.

I'm also trying to fix the confusing qqq.json documentation.

Bug: T368249
Depends-On: I9afbea452652c03adefa474b5ee5385e55d62af1
Change-Id: Id7a3df906064eaedc87a9ef56be170424b27281e